Added LogFormLabel to shared styledComponents
