+++ src/components/Layouts/Primary/TopBar/Items/AddMenu/Items/AddDonation/AddDonation.tsx
@@ -18,6 +18,7 @@ import * as yup from 'yup';
 import { AppealAutocomplete } from 'src/common/Autocompletes/AppealAutocomplete';
 import { DesignationAccountAutocomplete } from 'src/common/Autocompletes/DesignationAccountAutocomplete';
 import { useGetDesignationAccountsQuery } from 'src/components/EditDonationModal/EditDonationModal.generated';
+import { LogFormLabel } from 'src/components/Shared/styledComponents/styledComponents';
 import { CurrencyAutocomplete } from 'src/components/common/Autocomplete/CurrencyAutocomplete/CurrencyAutocomplete';
 import { DonorAccountAutocomplete } from 'src/components/common/Autocomplete/DonorAccountAutocomplete/DonorAccountAutocomplete';
 import { CustomDateField } from 'src/components/common/DateTimePickers/CustomDateField';
@@ -31,7 +32,7 @@ import {
   useAddDonationMutation,
   useGetDonationModalQuery,
 } from './AddDonation.generated';
-import { FormTextField, LogFormLabel } from './StyledComponents';
+import { FormTextField } from './StyledComponents';
 
 interface AddDonationProps {
   accountListId: string;
