+++ src/components/Layouts/Primary/TopBar/Items/AddMenu/Items/AddDonation/StyledComponents.tsx
@@ -1,14 +1,5 @@
-import { FormLabel, TextField } from '@mui/material';
+import { TextField } from '@mui/material';
 import { styled } from '@mui/material/styles';
-
-export const LogFormLabel = styled(FormLabel)(({ theme }) => ({
-  margin: theme.spacing(1, 0),
-  fontWeight: 'bold',
-  color: theme.palette.primary.dark,
-  '& span': {
-    color: theme.palette.error.main,
-  },
-}));
 
 export const FormTextField = styled(TextField)(({ theme }) => ({
   '& .MuiInputBase-root.Mui-disabled': {
